+++ CHANGELOG.md
@@ -5,6 +5,15 @@
   - Misc
 
 ### Unreleased
+
+  - Features Added
+    - ..
+  - Bugs Fixed
+    - ..
+
+
+### 2019-09-12 [actually two releses]
+
   - Features Added
     - Tweaks to the wallet provider UI
     - Allow to RedeemFromToken by passing a private key in the URL
@@ -16,6 +25,15 @@
     - fix crash in display of proposals' redeemables when account is readonly
     - show proposals as failing when the vote is tied
     - improvements, bug fixes on the Scheme Information page
+    - DAO Discussion page
+    - Add link to help center
+  - Bugs Fixed
+    - Fix number formatting
+    - Fix crashing of scheme page
+    - Hide notifications after 10 seconds
+    - Add dao address as default value for Cross-DAO redemptions
+    - Fix redemptions count
+
 
 ### version 0.9.4; 2019-08-13